@istio-testing 2021-12-02
@howardjohn: The following tests **failed**, say `/retest` to rerun all failed tests or `/retest-required` to rerun all mandatory failed tests:

Test name | Commit | Details | Required | Rerun command
--- | --- | --- | --- | ---
integ-pilot-istiodremote-tests_istio | 7869d8014493deeb5b24645c6628f0553ab504a3 | [link](https://prow.istio.io/view/gs/istio-prow/pr-logs/pull/istio_istio/36329/integ-pilot-istiodremote-tests_istio/1466477280258691072) | true | `/test integ-pilot-istiodremote-tests_istio`
integ-telemetry-k8s-tests_istio | 7869d8014493deeb5b24645c6628f0553ab504a3 | [link](https://prow.istio.io/view/gs/istio-prow/pr-logs/pull/istio_istio/36329/integ-telemetry-k8s-tests_istio/1466477278677438464) | true | `/test integ-telemetry-k8s-tests_istio`

<details>

Instructions for interacting with me using PR comments are available [here](https://git.k8s.io/community/contributors/guide/pull-requests.md).  If you have questions or suggestions related to my behavior, please file an issue against the [kubernetes/test-infra](https://github.com/kubernetes/test-infra/issues/new?title=Prow%20issue:) repository. I understand the commands that are listed [here](https://go.k8s.io/bot-commands).
</details>
<!-- test report -->